Variance matrix for a regional_mix object.

Calculates variance-covariance matrix from a regional_mix object

# S3 method for regional_mix
vcov(
  object,
  ...,
  object2 = NULL,
  method = "FiniteDifference",
  nboot = 1000,
  mc.cores = 1,
  D.accuracy = 2
)

Arguments

object
an object obtained from fitting a RCP (for region of common profile) mixture model. Such as that generated from a call to regional_mix(qv).

object2
an object of class regional_mix containing bootstrap samples of the parameter estimates (see regional_mix_boot(qv)). If NULL (default) the bootstrapping is performed from within the vcov function. If not null, then the vcov estimate is obtained from these bootstrap samples.

method
the method to calculate the variance-covariance matrix. Options are:'FiniteDifference' (default), BayesBoot, SimpleBoot, and EmpiricalInfo. The two bootstrap methods (BayesBoot and SimpleBoot, see regional_mix_boot(qv)) should be more general and may possibly be more robust. The EmpiricalInfo method implements an empirical estimate of the Fisher information matrix, I can not recommend it however. It seems to behave poorly, even in well behaved simulations. It is computationally thrifty though.

nboot
the number of bootstrap samples to take for the bootstrap estimation. Argument is ignored if !method %in% c(FiniteDifference,'EmpiricalInfo').

mc.cores
the number of cores to distrbute the calculations on. Default is 4. Set to 1 if the computer is running Windows (as it cannot handle forking -- see mclapply(qv)). Ignored if method=='EmpiricalInfo'.

D.accuracy
The number of finite difference points used in the numerical approximation to the observed information matrix. Ignored if method != FiniteDifference. Options are 2 (default) and 4.

\dots
Other calls to the vcov function.

Value

A square matrix of size equal to the number of parameters. It contains the variance matrix of the parameter estimates.

Details

If method is FiniteDifference, then the estimates variance matrix is based on a finite difference approximation to the observed information matrix. If method is either "BayesBoot" or "SimpleBoot", then the estimated variance matrix is calculated from bootstrap samples of the parameter estimates. See Foster et al (in prep) for details of how the bootstrapping is actually done, and regional_mix_boot(qv) for its implementation.

References

Foster, S.D., Givens, G.H., Dornan, G.J., Dunstan, P.K. and Darnell, R. (2013) Modelling Regions of Common Profiles Using Biological and Environmental Data. Environmetrics 24: 489--499. DOI: 10.1002/env.2245 Foster, S.D., Hill, N.A. and Lyons, M., 2017. Ecological grouping of survey sites when sampling artefacts are present. Journal of the Royal Statistical Society: Series C (Applied Statistics), 66(5), pp.1031-1047.