Estimate the Variance-covariance matrix for a species_mix object.

Calculates variance-covariance matrix from a species_mix object

# S3 method for species_mix
vcov(
  object,
  object2 = NULL,
  method = "BayesBoot",
  nboot = 10,
  mc.cores = 1,
  ...
)

Arguments

object
an object obtained from fitting a RCP (for region of common profile) mixture model. Such as that generated from a call to species_mix(qv).

object2
an object of class species_mix containing bootstrap samples of the parameter estimates (see species_mix_boot(qv)). If NULL (default) the bootstrapping is performed from within the vcov function. If not null, then the vcov estimate is obtained from these bootstrap samples.

method
the method to calculate the variance-covariance matrix. Options are:'FiniteDifference' (default), BayesBoot, SimpleBoot, and EmpiricalInfo. The two bootstrap methods (BayesBoot and SimpleBoot, see species_mix_boot(qv)) should be more general and may possibly be more robust. The EmpiricalInfo method implements an empirical estimate of the Fisher information matrix, I can not recommend it however. It seems to behave poorly, even in well behaved simulations. It is computationally thrifty though.

nboot
the number of bootstrap samples to take for the bootstrap estimation. Argument is ignored if !method %in% c(FiniteDifference,'EmpiricalInfo').

mc.cores
the number of cores to distribute the calculations on. Default is 4. Set to 1 if the computer is running Windows (as it cannot handle forking -- see mclapply(qv)). Ignored if method=='EmpiricalInfo'.

\dots
Ignored

Value

A square matrix of size equal to the number of parameters. It contains the variance matrix of the parameter estimates.

Details

If method is FiniteDifference, then the estimates variance matrix is based on a finite difference approximation to the observed information matrix. If method is either "BayesBoot" or "SimpleBoot", then the estimated variance matrix is calculated from bootstrap samples of the parameter estimates. See Foster et al (in prep) for details of how the bootstrapping is actually done, and species_mix_boot(qv) for its implementation.

Examples

# Estimate the variance-covariance matrix.
# This will provide estimates of uncertainty for model parameters.
# \donttest{
library(ecomix)
set.seed(42)
sam_form <- stats::as.formula(paste0('cbind(',paste(paste0('spp',1:20),
collapse = ','),")~x1+x2"))
sp_form <- ~ 1
beta <- matrix(c(-2.9,-3.6,-0.9,1,.9,1.9),3,2,byrow=TRUE)
dat <- data.frame(y=rep(1,100),x1=stats::runif(100,0,2.5),
x2=stats::rnorm(100,0,2.5))
dat[,-1] <- scale(dat[,-1])
simulated_data <- species_mix.simulate(archetype_formula = sam_form,species_formula = sp_form,
data = dat,beta=beta,family="bernoulli")

#> Random alpha from normal (-1,0.5) distribution

fm1 <- species_mix(archetype_formula = sam_form,species_formula = sp_form,
data = simulated_data, family = 'bernoulli',  nArchetypes=3)

#> SAM modelling

#> There are 3 archetypes to group the species into

#> There are 100 site observations for 20 species

#> The model for the archetype (grouping) is ~x1 + x2

#> The model for the species is ~1

#> You are implementing a bernoulli Species Archetype Model.

#> Using ECM algorithm to find starting values; using 1refits

#> ECM restart 1 of 1

#> Initialising starting values

#> Initial groups parameter estimates by K-means clustering with random noise

#> Iteration: 1| New loglik -967.597 | Ratio loglik 0

#> Iteration: 2| New loglik -841.415 | Ratio loglik 0.869592

#> Iteration: 3| New loglik -841.416 | Ratio loglik 1

#> initial  value 841.774554 
#> iter  10 value 820.166960
#> iter  20 value 819.991704
#> final  value 819.991378 
#> converged

vcov(fm1)# }

#> ..........

#>                        spp1          spp2          spp3          spp4
#> spp1           4.565090e-04  0.0020200790 -1.243787e-05  2.662469e-04
#> spp2           2.020079e-03  0.0697013946 -9.632371e-04 -2.809707e-04
#> spp3          -1.243787e-05 -0.0009632371  6.209984e-04 -1.121635e-05
#> spp4           2.662469e-04 -0.0002809707 -1.121635e-05  8.816217e-04
#> spp5           8.629653e-04  0.0516370829  1.306640e-04 -1.318055e-03
#> spp6           1.568467e-03  0.0503016099 -6.709463e-04 -3.180407e-04
#> spp7          -9.249576e-05  0.0010088902  4.167081e-04 -8.191552e-05
#> spp8           3.034812e-04 -0.0013506744 -1.649223e-04  7.327230e-04
#> spp9          -4.687352e-04  0.0020733860  9.701559e-04 -5.645067e-04
#> spp10         -1.698162e-05  0.0002706752  1.002621e-04 -1.003380e-05
#> spp11          2.025462e-04 -0.0013710275 -1.032829e-04  5.369576e-04
#> spp12          3.266690e-04 -0.0015335049 -1.705131e-04  8.115590e-04
#> spp13          1.078767e-03  0.0288389019 -4.182052e-04 -4.142962e-04
#> spp14          1.161544e-03  0.0303574072 -3.967581e-04 -5.405848e-04
#> spp15         -1.838351e-04 -0.0052714043 -1.069436e-04  8.791660e-04
#> spp16          1.209823e-04 -0.0005006032 -6.015833e-05  2.850100e-04
#> spp17          5.930300e-04  0.0061900421  9.567067e-04  7.062637e-04
#> spp18         -1.572932e-03  0.0005228968  8.716349e-04 -2.192168e-03
#> spp19          2.424640e-04 -0.0009238257 -1.117998e-04  6.556029e-04
#> spp20          1.514718e-04 -0.0009133881 -4.375979e-05  3.533701e-04
#> x1.Archetype1  2.183143e-03  0.1153490900 -1.524163e-03  5.724138e-04
#> x1.Archetype2  5.603767e-04 -0.0005320476 -9.325333e-05  1.409392e-03
#> x1.Archetype3  8.989354e-04 -0.0039148712 -1.927434e-03  8.706752e-04
#> x2.Archetype1  4.892889e-03  0.1371967559 -1.832567e-03 -1.250322e-03
#> x2.Archetype2 -6.225131e-04  0.0050160722  5.364194e-04 -1.525393e-03
#> x2.Archetype3 -1.847842e-04 -0.0035016540 -7.636314e-04 -1.449791e-04
#> eta1           0.000000e+00  0.0000000000  0.000000e+00  0.000000e+00
#> eta2           0.000000e+00  0.0000000000  0.000000e+00  0.000000e+00
#>                        spp5          spp6          spp7          spp8
#> spp1           0.0008629653  0.0015684667 -9.249576e-05  3.034812e-04
#> spp2           0.0516370829  0.0503016099  1.008890e-03 -1.350674e-03
#> spp3           0.0001306640 -0.0006709463  4.167081e-04 -1.649223e-04
#> spp4          -0.0013180555 -0.0003180407 -8.191552e-05  7.327230e-04
#> spp5           0.0468466898  0.0370923695  1.714678e-03 -2.097463e-03
#> spp6           0.0370923695  0.0364142240  6.899888e-04 -1.018461e-03
#> spp7           0.0017146783  0.0006899888  4.512966e-04 -2.262818e-04
#> spp8          -0.0020974635 -0.0010184613 -2.262818e-04  7.777180e-04
#> spp9           0.0034950087  0.0014060110  1.075370e-03 -9.828176e-04
#> spp10          0.0003781463  0.0001860982  1.072509e-04 -5.507934e-05
#> spp11         -0.0017591135 -0.0010219213 -1.455511e-04  5.754325e-04
#> spp12         -0.0023450442 -0.0011551949 -2.383236e-04  8.613411e-04
#> spp13          0.0211385249  0.0210643482  3.017429e-04 -6.432362e-04
#> spp14          0.0219462666  0.0222907374  4.119444e-04 -6.649557e-04
#> spp15         -0.0035173189 -0.0040604930  1.361834e-05  6.499726e-04
#> spp16         -0.0007625162 -0.0003771202 -8.511419e-05  3.050897e-04
#> spp17          0.0048658609  0.0041750555  1.244196e-03  3.829810e-04
#> spp18         -0.0003899478  0.0002271116  1.303621e-03 -2.962070e-03
#> spp19         -0.0017051716 -0.0007237658 -1.756243e-04  7.065216e-04
#> spp20         -0.0010034093 -0.0006801101 -7.806481e-05  3.882862e-04
#> x1.Archetype1  0.0885968621  0.0820077270  2.170639e-03 -1.991205e-03
#> x1.Archetype2 -0.0022108890 -0.0004968110 -8.031454e-05  1.450619e-03
#> x1.Archetype3 -0.0055100983 -0.0025797359 -2.233348e-03  1.937765e-03
#> x2.Archetype1  0.0996798122  0.0999389483  1.653498e-03 -2.741058e-03
#> x2.Archetype2  0.0062486587  0.0036901083  7.834419e-04 -1.662391e-03
#> x2.Archetype3 -0.0055979427 -0.0025144425 -7.205131e-04 -1.229767e-04
#> eta1           0.0000000000  0.0000000000  0.000000e+00  0.000000e+00
#> eta2           0.0000000000  0.0000000000  0.000000e+00  0.000000e+00
#>                        spp9         spp10         spp11         spp12
#> spp1          -0.0004687352 -1.698162e-05  2.025462e-04  3.266690e-04
#> spp2           0.0020733860  2.706752e-04 -1.371027e-03 -1.533505e-03
#> spp3           0.0009701559  1.002621e-04 -1.032829e-04 -1.705131e-04
#> spp4          -0.0005645067 -1.003380e-05  5.369576e-04  8.115590e-04
#> spp5           0.0034950087  3.781463e-04 -1.759114e-03 -2.345044e-03
#> spp6           0.0014060110  1.860982e-04 -1.021921e-03 -1.155195e-03
#> spp7           0.0010753703  1.072509e-04 -1.455511e-04 -2.383236e-04
#> spp8          -0.0009828176 -5.507934e-05  5.754325e-04  8.613411e-04
#> spp9           0.0030545882  2.741672e-04 -6.761006e-04 -1.069965e-03
#> spp10          0.0002741672  2.869498e-05 -3.460082e-05 -6.045532e-05
#> spp11         -0.0006761006 -3.460082e-05  4.332642e-04  6.374488e-04
#> spp12         -0.0010699654 -6.045532e-05  6.374488e-04  9.571245e-04
#> spp13          0.0005392221  7.874509e-05 -6.290863e-04 -7.273864e-04
#> spp14          0.0006749438  8.907226e-05 -6.328931e-04 -7.260863e-04
#> spp15         -0.0002130568  2.645406e-05  5.473996e-04  7.142649e-04
#> spp16         -0.0003874520 -2.185691e-05  2.254553e-04  3.387095e-04
#> spp17          0.0025343838  3.259726e-04  3.049785e-04  4.069052e-04
#> spp18          0.0061070194  4.607346e-04 -2.095366e-03 -3.345189e-03
#> spp19         -0.0008486013 -4.934367e-05  5.197706e-04  7.885398e-04
#> spp20         -0.0004587978 -2.594902e-05  2.871477e-04  4.355145e-04
#> x1.Archetype1  0.0046559473  5.711819e-04 -2.107101e-03 -2.288370e-03
#> x1.Archetype2 -0.0012212504 -4.684672e-05  1.075109e-03  1.642676e-03
#> x1.Archetype3 -0.0067824270 -6.547140e-04  1.279451e-03  2.173236e-03
#> x2.Archetype1  0.0030999077  4.449683e-04 -2.740671e-03 -3.092108e-03
#> x2.Archetype2  0.0026348864  1.683540e-04 -1.230971e-03 -1.811819e-03
#> x2.Archetype3 -0.0007474438 -8.365871e-05 -7.739024e-05 -2.182442e-04
#> eta1           0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00
#> eta2           0.0000000000  0.000000e+00  0.000000e+00  0.000000e+00
#>                       spp13         spp14         spp15         spp16
#> spp1           1.078767e-03  1.161544e-03 -1.838351e-04  1.209823e-04
#> spp2           2.883890e-02  3.035741e-02 -5.271404e-03 -5.006032e-04
#> spp3          -4.182052e-04 -3.967581e-04 -1.069436e-04 -6.015833e-05
#> spp4          -4.142962e-04 -5.405848e-04  8.791660e-04  2.850100e-04
#> spp5           2.113852e-02  2.194627e-02 -3.517319e-03 -7.625162e-04
#> spp6           2.106435e-02  2.229074e-02 -4.060493e-03 -3.771202e-04
#> spp7           3.017429e-04  4.119444e-04  1.361834e-05 -8.511419e-05
#> spp8          -6.432362e-04 -6.649557e-04  6.499726e-04  3.050897e-04
#> spp9           5.392221e-04  6.749438e-04 -2.130568e-04 -3.874520e-04
#> spp10          7.874509e-05  8.907226e-05  2.645406e-05 -2.185691e-05
#> spp11         -6.290863e-04 -6.328931e-04  5.473996e-04  2.254553e-04
#> spp12         -7.273864e-04 -7.260863e-04  7.142649e-04  3.387095e-04
#> spp13          1.252194e-02  1.346969e-02 -2.747513e-03 -2.357019e-04
#> spp14          1.346969e-02  1.495441e-02 -3.078700e-03 -2.372184e-04
#> spp15         -2.747513e-03 -3.078700e-03  3.072962e-03  2.483437e-04
#> spp16         -2.357019e-04 -2.372184e-04  2.483437e-04  1.202411e-04
#> spp17          1.854454e-03  1.795363e-03 -6.325379e-04  1.406293e-04
#> spp18         -4.627448e-04 -9.418625e-04 -1.352307e-03 -1.235554e-03
#> spp19         -4.997870e-04 -4.824303e-04  2.903900e-04  2.786859e-04
#> spp20         -4.074294e-04 -3.898756e-04  3.502634e-04  1.551477e-04
#> x1.Archetype1  4.493003e-02  4.561543e-02 -6.846575e-03 -7.445770e-04
#> x1.Archetype2 -4.301283e-04 -8.689500e-05  1.020220e-03  5.789704e-04
#> x1.Archetype3 -7.278720e-04 -6.106183e-04 -4.060212e-04  7.969317e-04
#> x2.Archetype1  5.886137e-02  6.322905e-02 -1.194024e-02 -1.004443e-03
#> x2.Archetype2  2.218142e-03  2.594527e-03 -1.596534e-03 -6.436995e-04
#> x2.Archetype3 -1.495338e-03 -2.127202e-03  3.998539e-04 -9.310884e-05
#> eta1           0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#> eta2           0.000000e+00  0.000000e+00  0.000000e+00  0.000000e+00
#>                       spp17         spp18         spp19         spp20
#> spp1           0.0005930300 -0.0015729317  2.424640e-04  1.514718e-04
#> spp2           0.0061900421  0.0005228968 -9.238257e-04 -9.133881e-04
#> spp3           0.0009567067  0.0008716349 -1.117998e-04 -4.375979e-05
#> spp4           0.0007062637 -0.0021921683  6.556029e-04  3.533701e-04
#> spp5           0.0048658609 -0.0003899478 -1.705172e-03 -1.003409e-03
#> spp6           0.0041750555  0.0002271116 -7.237658e-04 -6.801101e-04
#> spp7           0.0012441958  0.0013036213 -1.756243e-04 -7.806481e-05
#> spp8           0.0003829810 -0.0029620704  7.065216e-04  3.882862e-04
#> spp9           0.0025343838  0.0061070194 -8.486013e-04 -4.587978e-04
#> spp10          0.0003259726  0.0004607346 -4.934367e-05 -2.594902e-05
#> spp11          0.0003049785 -0.0020953663  5.197706e-04  2.871477e-04
#> spp12          0.0004069052 -0.0033451893  7.885398e-04  4.355145e-04
#> spp13          0.0018544544 -0.0004627448 -4.997870e-04 -4.074294e-04
#> spp14          0.0017953634 -0.0009418625 -4.824303e-04 -3.898756e-04
#> spp15         -0.0006325379 -0.0013523071  2.903900e-04  3.502634e-04
#> spp16          0.0001406293 -0.0012355539  2.786859e-04  1.551477e-04
#> spp17          0.0106391891  0.0038391829  5.729962e-04  2.508475e-04
#> spp18          0.0038391829  0.0238905543 -2.755603e-03 -1.774307e-03
#> spp19          0.0005729962 -0.0027556029  7.196987e-04  3.510373e-04
#> spp20          0.0002508475 -0.0017743068  3.510373e-04  2.180233e-04
#> x1.Archetype1  0.0142426221  0.0028738504 -1.008726e-03 -1.407358e-03
#> x1.Archetype2  0.0019556712 -0.0058125227  1.410183e-03  7.983237e-04
#> x1.Archetype3 -0.0067582111 -0.0168580456  1.829129e-03  1.079690e-03
#> x2.Archetype1  0.0093461410 -0.0014015228 -2.050769e-03 -1.822826e-03
#> x2.Archetype2  0.0003685530  0.0060946490 -1.431544e-03 -7.681361e-04
#> x2.Archetype3 -0.0014844273  0.0050339257 -3.123683e-04 -3.170312e-04
#> eta1           0.0000000000  0.0000000000  0.000000e+00  0.000000e+00
#> eta2           0.0000000000  0.0000000000  0.000000e+00  0.000000e+00
#>               x1.Archetype1 x1.Archetype2 x1.Archetype3 x2.Archetype1
#> spp1           0.0021831430  5.603767e-04  0.0008989354  0.0048928890
#> spp2           0.1153490900 -5.320476e-04 -0.0039148712  0.1371967559
#> spp3          -0.0015241631 -9.325333e-05 -0.0019274338 -0.0018325674
#> spp4           0.0005724138  1.409392e-03  0.0008706752 -0.0012503221
#> spp5           0.0885968621 -2.210889e-03 -0.0055100983  0.0996798122
#> spp6           0.0820077270 -4.968110e-04 -0.0025797359  0.0999389483
#> spp7           0.0021706386 -8.031454e-05 -0.0022333478  0.0016534983
#> spp8          -0.0019912052  1.450619e-03  0.0019377647 -0.0027410584
#> spp9           0.0046559473 -1.221250e-03 -0.0067824270  0.0030999077
#> spp10          0.0005711819 -4.684672e-05 -0.0006547140  0.0004449683
#> spp11         -0.0021071010  1.075109e-03  0.0012794512 -0.0027406709
#> spp12         -0.0022883698  1.642676e-03  0.0021732360 -0.0030921081
#> spp13          0.0449300339 -4.301283e-04 -0.0007278720  0.0588613745
#> spp14          0.0456154334 -8.689500e-05 -0.0006106183  0.0632290514
#> spp15         -0.0068465746  1.020220e-03 -0.0004060212 -0.0119402397
#> spp16         -0.0007445770  5.789704e-04  0.0007969317 -0.0010044427
#> spp17          0.0142426221  1.955671e-03 -0.0067582111  0.0093461410
#> spp18          0.0028738504 -5.812523e-03 -0.0168580456 -0.0014015228
#> spp19         -0.0010087260  1.410183e-03  0.0018291286 -0.0020507693
#> spp20         -0.0014073584  7.983237e-04  0.0010796898 -0.0018228256
#> x1.Archetype1  0.2055185573 -3.281387e-04 -0.0095741396  0.2162938836
#> x1.Archetype2 -0.0003281387  3.369448e-03  0.0029570257 -0.0012452917
#> x1.Archetype3 -0.0095741396  2.957026e-03  0.0175047680 -0.0050906679
#> x2.Archetype1  0.2162938836 -1.245292e-03 -0.0050906679  0.2781091789
#> x2.Archetype2  0.0078434524 -2.526558e-03 -0.0047507875  0.0100273459
#> x2.Archetype3 -0.0061146510 -1.513915e-03 -0.0009445918 -0.0070510520
#> eta1           0.0000000000  0.000000e+00  0.0000000000  0.0000000000
#> eta2           0.0000000000  0.000000e+00  0.0000000000  0.0000000000
#>               x2.Archetype2 x2.Archetype3 eta1 eta2
#> spp1          -0.0006225131 -1.847842e-04    0    0
#> spp2           0.0050160722 -3.501654e-03    0    0
#> spp3           0.0005364194 -7.636314e-04    0    0
#> spp4          -0.0015253933 -1.449791e-04    0    0
#> spp5           0.0062486587 -5.597943e-03    0    0
#> spp6           0.0036901083 -2.514443e-03    0    0
#> spp7           0.0007834419 -7.205131e-04    0    0
#> spp8          -0.0016623907 -1.229767e-04    0    0
#> spp9           0.0026348864 -7.474438e-04    0    0
#> spp10          0.0001683540 -8.365871e-05    0    0
#> spp11         -0.0012309713 -7.739024e-05    0    0
#> spp12         -0.0018118194 -2.182442e-04    0    0
#> spp13          0.0022181415 -1.495338e-03    0    0
#> spp14          0.0025945270 -2.127202e-03    0    0
#> spp15         -0.0015965345  3.998539e-04    0    0
#> spp16         -0.0006436995 -9.310884e-05    0    0
#> spp17          0.0003685530 -1.484427e-03    0    0
#> spp18          0.0060946490  5.033926e-03    0    0
#> spp19         -0.0014315436 -3.123683e-04    0    0
#> spp20         -0.0007681361 -3.170312e-04    0    0
#> x1.Archetype1  0.0078434524 -6.114651e-03    0    0
#> x1.Archetype2 -0.0025265580 -1.513915e-03    0    0
#> x1.Archetype3 -0.0047507875 -9.445918e-04    0    0
#> x2.Archetype1  0.0100273459 -7.051052e-03    0    0
#> x2.Archetype2  0.0040544335 -8.470673e-04    0    0
#> x2.Archetype3 -0.0008470673  4.353407e-03    0    0
#> eta1           0.0000000000  0.000000e+00    0    0
#> eta2           0.0000000000  0.000000e+00    0    0