data_preperation-class

prepares table of observational records into a sites x species matrix for use in ecomix.

table_to_species_data(observation_data, site_id = "site_id",
  species_id = "species_id", distribution = "bernoulli",
  measurement_id = NULL)

make_mixture_data(species_data, covariate_data)

Arguments

observation_data
data to be imported into R. A table of species occurrences and site records.

site_id
Name of column that contains site names /IDs.

species_id
Name of column that contains species information/ID.

distribution
The family of statistical distribution to use within the ecomix models. A choice between "bernoulli", "poisson", "ippm", "negative_binomial", "tweedie" and "gaussian" distributions are possible and applicable to specific types of data. The data perp will try to match distribution with the data.

measurement_id
Name of column that contains species observational data. eg, 'occurrence', 'abundance' or 'biomass'.

species_data
A character string providing the name of the column in data that contains the counts

covariate_data
A character string providing the name of the column in data that contains the binary occurence data

Value

A sites x species matrix to used as species_data in ecomix models.

A list containing the following elements:

species_data

A matrix of occurrence, counts, biomass, or...

covariate_data

The design matrix for the global covariates